Moved conceptandrelationview events to correct package
