+++ web-ui/src/main/java/org/archcnl/ui/common/conceptandrelationlistview/HierarchyView.java
@@ -11,8 +11,8 @@ import java.util.ArrayList;
 import java.util.List;
 import org.archcnl.domain.common.HierarchyNode;
 import org.archcnl.domain.common.conceptsandrelations.HierarchyObject;
-import org.archcnl.ui.events.GridUpdateRequestedEvent;
-import org.archcnl.ui.events.HierarchySwapRequestedEvent;
+import org.archcnl.ui.common.conceptandrelationlistview.events.GridUpdateRequestedEvent;
+import org.archcnl.ui.common.conceptandrelationlistview.events.HierarchySwapRequestedEvent;
 import org.archcnl.ui.inputview.rulesormappingeditorview.RulesOrMappingEditorView;
 
 public class HierarchyView<T extends HierarchyObject> extends RulesOrMappingEditorView {
